Linear Regression ML Example for TensorFlow Lite
This is an unineural network. There is only one neuron.

In this case, the optimizations for the Edge TPU won't be practical. In fact, average execution time will be increased, due to additional steps on quantization and dequantization.

Nevertheless, it also is an example about how to perform the optimizations.